+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add support for the new network configuration format introduced in pot 0.8
+
+### Changed
+- Adopted the crate IpNet to perform operations on IP addresses
+- Add IPv6 support everywhere
 
 ## [0.1.3] 2019-04-03
 ### Added