+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] 2019-08-02
 ### Added
 - New utility, called potcpu, designed to manage cpusets for pot
 - potcpu: rebalance command, to distribute allocation all over the CPUs